code cleanup after moving into appCode

+++ iBeaconHarvester/MenuTableViewController.m
@@ -11,18 +11,17 @@
 #import "MenuTableCell.h"
 
 @interface MenuTableViewController ()
-@property (nonatomic, strong) NSArray *menuItems;
+@property(nonatomic, strong) NSArray *menuItems;
 @end
 
 @implementation MenuTableViewController
 
-- (void)viewDidLoad
-{
+- (void)viewDidLoad {
     [super viewDidLoad];
-    
+
     SWRevealViewController *revealController = self.revealViewController;
     [self.view addGestureRecognizer:revealController.panGestureRecognizer];
-    
+
     self.view.backgroundColor = [UIColor colorWithWhite:0.2f alpha:1.0f];
     self.tableView.backgroundColor = [UIColor colorWithWhite:0.2f alpha:1.0f];
     self.tableView.separatorColor = [UIColor colorWithWhite:0.8f alpha:0.4f];
@@ -31,19 +30,16 @@
 
 #pragma mark - Table view data source
 
-- (NSInteger)numberOfSectionsInTableView:(UITableView *)tableView
-{
+- (NSInteger)numberOfSectionsInTableView:(UITableView *)tableView {
     return 1;
 }
 
-- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section
-{
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
     return [self.menuItems count];
 }
 
-- (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath
-{
-    NSString *cellIdentifier = [self.menuItems objectAtIndex:indexPath.row];
+- (UITableViewCell *)tableView:(UITableView *)tableView cellForRowAtIndexPath:(NSIndexPath *)indexPath {
+    NSString *cellIdentifier = [self.menuItems objectAtIndex:(NSUInteger) indexPath.row];
     MenuTableCell *cell = [self.tableView dequeueReusableCellWithIdentifier:cellIdentifier];
     if (cell == nil) {
         cell = [[MenuTableCell alloc] initWithStyle:UITableViewCellStyleSubtitle reuseIdentifier:cellIdentifier];
@@ -55,28 +51,24 @@
 
 
 // Override to support conditional editing of the table view.
-- (BOOL)tableView:(UITableView *)tableView canEditRowAtIndexPath:(NSIndexPath *)indexPath
-{
+- (BOOL)tableView:(UITableView *)tableView canEditRowAtIndexPath:(NSIndexPath *)indexPath {
     return NO;
 }
 
 #pragma mark - Seque handling
 
-- (void) prepareForSegue: (UIStoryboardSegue *) segue sender: (id) sender
-{ 
+- (void)prepareForSegue:(UIStoryboardSegue *)segue sender:(id)sender {
     // configure the segue.
-    if ( [segue isKindOfClass: [SWRevealViewControllerSegue class]] )
-    {
-        SWRevealViewControllerSegue* rvcs = (SWRevealViewControllerSegue*) segue;
-        
-        SWRevealViewController* rvc = self.revealViewController;
+    if ([segue isKindOfClass:[SWRevealViewControllerSegue class]]) {
+        SWRevealViewControllerSegue *rvcs = (SWRevealViewControllerSegue *) segue;
+
+        SWRevealViewController *rvc = self.revealViewController;
         NSAssert( rvc != nil, @"oops! must have a revealViewController" );
-        
-        NSAssert( [rvc.frontViewController isKindOfClass: [UINavigationController class]], @"oops!  for this segue we want a permanent navigation controller in the front!" );
-        
-        rvcs.performBlock = ^(SWRevealViewControllerSegue* rvc_segue, UIViewController* svc, UIViewController* dvc)
-        {
-            UINavigationController* nc = [[UINavigationController alloc] initWithRootViewController:dvc];
+
+        NSAssert( [rvc.frontViewController isKindOfClass:[UINavigationController class]], @"oops!  for this segue we want a permanent navigation controller in the front!" );
+
+        rvcs.performBlock = ^(SWRevealViewControllerSegue *rvc_segue, UIViewController *svc, UIViewController *dvc) {
+            UINavigationController *nc = [[UINavigationController alloc] initWithRootViewController:dvc];
             [rvc setFrontViewController:nc animated:YES];
         };
     }
